Type에서 선언한 변수명으로 변경, 클라이언트 타입 캐스팅 Warning 메시지 수정

+++ Client/Client.cpp
@@ -43,7 +43,7 @@ public:
         std::memset(packet.header.checkSum, 0x12, sizeof(packet.header.checkSum));
         std::memset(packet.header.checkSum + 1, 0x34, sizeof(packet.header.checkSum) - 1);
         packet.header.type = PacketType::YJ;
-        packet.header.size = sizeof(PacketBuffer) + message.size();
+        packet.header.size = static_cast<uint32>(sizeof(PacketBuffer) + message.size());
         std::memcpy(packet.payload, message.c_str(), message.size());
         packet.tail.value = 255;
